More Troop Work
Troops now attack with a delay, Troops now can only spawn after 4
seconds for swordsman, 6 for archer etc, these values are currently
hardcoded for the troop types, that should be changed to use the troops
actual value somehow, possibly by changing the troop class setup to be a
troop parent class and swordsman, archer, etc child classes. Changes the
spawn position from outside the base to inside due to a crash. Rotated
enemy troops to be facing the correct way, changed the text on screen to
be togglable, both player and enemy troops now attack properly and die
properly