Troop Spawn Change & Upgrades
Changes the places troops spawn, so they spawn in a range area on
different z layers, enemy and player troops can spawn in the same z area
range also.
Upgrades for troops and base also implemented, not fully, need to add
cost to the upgrades to finish off for base and troops.